+++ src/lib/tuning/equaltemperament.cc
@@ -3,7 +3,7 @@
 
 Equal_temperament::Equal_temperament
 (
-    float steps_per_interval,
+    int steps_per_interval,
     float interval_size,
     float base_frequency
 )
@@ -19,12 +19,12 @@ float Equal_temperament::frequency_at(float steps) const
     return m_base_frequency * pow(step_multiplier(), steps);
 }
 
-float Equal_temperament::steps_per_interval() const
+int Equal_temperament::steps_per_interval() const
 {
     return m_steps_per_interval;
 }
 
-void Equal_temperament::set_steps_per_interval(float value)
+void Equal_temperament::set_steps_per_interval(int value)
 {
     m_steps_per_interval = value;
 }
